prefetch content tests and cleanup
